SpriteCache: read old spritefile topmost index as unsigned 16-bit
This lets to read old format files which contain up to 64k sprites; even though these would not be usable in practice before due to 30k sprite limit.

+++ Common/ac/spritecache.cpp
@@ -815,7 +815,7 @@ HError SpriteCache::InitFile(const char *filename, const char *sprindex_filename
 
     sprkey_t topmost;
     if (vers < kSprfVersion_HighSpriteLimit)
-        topmost = _stream->ReadInt16();
+        topmost = (uint16_t)_stream->ReadInt16();
     else
         topmost = _stream->ReadInt32();
     if (vers < kSprfVersion_Uncompressed)
